新機能：他のユーザーのサウナログの閲覧、投稿したユーザーのみ編集・削除の認可

+++ backend/app/Http/Controllers/SaunalogController.php
@@ -22,6 +22,14 @@ class SaunalogController extends Controller
         $this->middleware('auth');
         $this->saunalogService = $saunalogService;
     }
+    
+
+    //すべてのサウナログを一覧表示
+    public function getAllLogs() 
+    {
+        return response()->json($this->saunalogService->getAllLogs());
+    }
+
 
     //ユーザーに紐づくサウナログを一覧で表示させる
     public function getLogs(Request $request)
@@ -29,6 +37,7 @@ class SaunalogController extends Controller
         $userId = $request->user()->id;
         return $this->saunalogService->getLogsByUser($userId);
     }
+
 
     //特定のサウナログのみの表示
     public function getLogById($id)
@@ -41,6 +50,7 @@ class SaunalogController extends Controller
         }
     }
 
+
     //サウナログの新規作成
     public function create(CreateSaunalogRequest $request)
     {
@@ -48,9 +58,12 @@ class SaunalogController extends Controller
         return $this->saunalogService->createLog($request->validated(), $user);
     }
 
+
     //特定のサウナログの編集
     public function update($id, UpdateSaunalogRequest $request)
     {
+        $saunalog = Saunalog::findOrFail($id); // IDを使ってサウナログのインスタンスを取得
+        $this->authorize('update', $saunalog); 
         try {
             return $this->saunalogService->updateLogById($id, $request->validated());
         } catch(NotFoundException $e) {
@@ -58,9 +71,12 @@ class SaunalogController extends Controller
         }
     }
 
+
     //特定のサウナログの削除
     public function delete($id)
     {
+        $saunalog = Saunalog::findOrFail($id);
+        $this->authorize('delete', $saunalog);
         try {
             return $this->saunalogService->deleteLogById($id);
         } catch(NotFoundException $e) {
